# LRU generics