--- a/2q.go
+++ b/2q.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 	"sync"
 
-	"github.com/descope/golang-lru/simplelru"
+	"github.com/hashicorp/golang-lru/simplelru"
 )
 
 const (