--- a/go.mod
+++ b/go.mod
@@ -1,3 +1,5 @@
-module github.com/hashicorp/golang-lru
+module github.com/descope/golang-lru
 
-go 1.12
+go 1.18
+
+require github.com/bahlo/generic-list-go v0.2.0 // indirect