… 2q_test.go
… arc_test.go
… lru_test.go
… simplelru/lru_test.go