Added generics for all levels of LRU implementation
For internal list, moved from GO built in containers/list to a generic implementation:
github.com/bahlo/generic-list-go
Updated all tests to accommodate generics
In go.mod changed go version to 1.18