@jefferai 2022-09-14
I don't know who would make this decision, but just a comment: likely to avoid breaking compatibility with earlier users we'd likely want to wait until either Go 1.18 is unsupported or make a `v2` of the package.