@AskAlexSharov 2022-10-06
To avoid `v2` introduction - can follow `google/btree` approach to introducing generics:
https://github.com/google/btree/pull/45
leave existing types as is, and add new generic type:
```
 func (t *Btree)     Get(item Item) Item      { ... }
 func (t *BtreeG[T]) Get(item T)    (T, bool) { ... }
```