@jefferai 2022-10-11
@aviadl What are your thoughts on @AskAlexSharov 's suggestion. It makes sense to me as a good way to allow generics improvements but keep compatibility.