@aviadl 2022-10-12
@jefferai I think that working with go.mod would actually be better here
Code would look nicer for the long run, instead of inventing a new object type