@jefferai 2022-10-21
It seems like Google's `btree` has never been tagged so they probably didn't want to introduce `v2` in order to not have to start tagging it.

I think it's probably fine here to just go v2 and use the generic interface.

I'm still skeptical of the change to an outside container library; currently the library has no outside dependencies and I don't really want to add them. But I also don't know what the issue is.

@SoMuchForSubtlety Are there meaningful differences between your version and what's proposed here, other than the `container/list` question?