@AskAlexSharov 2022-10-22
> What's the issue with `container/list` in the first place?

it doesn't support generics