@SoMuchForSubtlety 2022-10-22
> @SoMuchForSubtlety Are there meaningful differences between your version and what's proposed here, other than the `container/list` question?

Not really. I kept the naked returns instead of explicitly declaring empty return values for misses, but that's just a style choice.

Here is the benchstat of my code and this PR
```
$ benchstat pr.txt mine.txt
name         old time/op    new time/op    delta
2Q_Rand-16      587ns ± 9%     516ns ±10%   -12.20%  (p=0.000 n=16+16)
2Q_Freq-16      494ns ±12%     442ns ± 6%   -10.51%  (p=0.000 n=20+17)
ARC_Rand-16     758ns ± 8%     683ns ±10%    -9.90%  (p=0.000 n=20+20)
ARC_Freq-16     619ns ±11%     549ns ±14%   -11.24%  (p=0.000 n=20+20)
LRU_Rand-16     240ns ± 8%     216ns ± 8%   -10.06%  (p=0.000 n=20+20)
LRU_Freq-16     220ns ± 5%     200ns ± 7%    -9.07%  (p=0.000 n=20+20)

name         old alloc/op   new alloc/op   delta
2Q_Rand-16      67.0B ± 0%     67.0B ± 0%      ~     (all equal)
2Q_Freq-16      59.0B ± 0%     59.0B ± 0%      ~     (all equal)
ARC_Rand-16     84.4B ± 1%     84.0B ± 0%    -0.47%  (p=0.002 n=20+19)
ARC_Freq-16     67.9B ± 5%     67.2B ± 6%      ~     (p=0.099 n=20+20)
LRU_Rand-16     36.0B ± 0%     36.0B ± 0%      ~     (all equal)
LRU_Freq-16     33.0B ± 0%     33.0B ± 0%      ~     (all equal)

name         old allocs/op  new allocs/op  delta
2Q_Rand-16       2.00 ± 0%      1.00 ± 0%   -50.00%  (p=0.000 n=20+20)
2Q_Freq-16       2.00 ± 0%      1.00 ± 0%   -50.00%  (p=0.000 n=20+20)
ARC_Rand-16      3.00 ± 0%      1.00 ± 0%   -66.67%  (p=0.000 n=20+20)
ARC_Freq-16      2.00 ± 0%      1.00 ± 0%   -50.00%  (p=0.000 n=20+20)
LRU_Rand-16      1.00 ± 0%      0.00       -100.00%  (p=0.000 n=20+20)
LRU_Freq-16      1.00 ± 0%      0.00       -100.00%  (p=0.000 n=20+20)
```