@jefferai 2022-10-24
>it doesn't support generics

Oh, I assumed it had been modified to do so since @SoMuchForSubtlety said he vendored it :-)