@jefferai 2022-10-24
>I kept the naked returns instead of explicitly declaring empty return values for misses, but that's just a style choice.

It isn't :-)  An empty object could have been explicitly added to the cache. You can't differentiate between the two if you return empty objects on misses. So your style is more correct. Are you planning on PRing it?